requirements.txt
discord.py
sphinx
light-koreanbots
pygithub
gitpython
beautifulsoup4